manage product created

+++ src/Pages/Dashboard/MakeAdmin.js
@@ -4,6 +4,7 @@ import { useQuery } from 'react-query';
 import { toast } from 'react-toastify';
 import Loading from '../Shared/Loading';
 import auth from '../../firebase.init';
+import swal from 'sweetalert';
 
 const Users = () => {
     const { data: users, isLoading, refetch } = useQuery('users', () => fetch(`${process.env.REACT_APP_SERVER_LINK}/user`, {
@@ -40,28 +41,35 @@ const Users = () => {
     }
 
     const deleteUser = (id) => {
-        fetch(`${process.env.REACT_APP_SERVER_LINK}/user/${id}`, {
-            method: 'DELETE',
-            headers: {
-                authorization: `Bearer ${localStorage.getItem('accessToken')}`
-            }
+        swal("Are you sure you want to delete your comment?", {
+            buttons: ["No!", true],
         })
-            .then(res => {
-                if (res.status === 401 || res.status === 403) {
-                    signOut(auth);
-                    localStorage.removeItem('accessToken');
+            .then(proceed => {
+                if (proceed) {
+                    fetch(`${process.env.REACT_APP_SERVER_LINK}/user/${id}`, {
+                        method: 'DELETE',
+                        headers: {
+                            authorization: `Bearer ${localStorage.getItem('accessToken')}`
+                        }
+                    })
+                        .then(res => {
+                            if (res.status === 401 || res.status === 403) {
+                                signOut(auth);
+                                localStorage.removeItem('accessToken');
+                            }
+                            return res.json()
+                        })
+                        .then(data => {
+                            refetch();
+                            toast("User delete");
+                        })
                 }
-                return res.json()
-            })
-            .then(data => {
-                refetch();
-                toast("User delete");
             })
     }
 
     return (
-        <div>
-            <h2 className="text-2xl">All Users: {users.length}</h2>
+        <div className='container mx-auto'>
+            <h2 className="text-2xl font-bold my-4">All Users: {users.length}</h2>
             <div className="overflow-x-auto">
                 <table className="table w-full">
                     <thead>
@@ -81,7 +89,7 @@ const Users = () => {
                                     :
                                     <div className="badge badge-accent">Admin</div>}
                                 </td>
-                                <td><button className="btn btn-xs" onClick={()=>deleteUser(user._id)}>Remove User</button></td>
+                                <td><button className="btn btn-xs" onClick={() => deleteUser(user._id)}>Remove User</button></td>
                             </tr>)
                         }
                     </tbody>
